requirements.txt
openai>=1.107.1
openai-agents>=0.3.3
atendentepro>=0.5.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0